Support slices in species arg, e.g. 1:30 means build species that are 1 mod 30
Ensure genes_info.txt is uploaded last, only on complete success.

Ensure some log is uploaded even if the command crashed before creating a log.

Ensure we start building as soon as we find a species that is not yet build,
insted of doing a pretask file checking pass first.